% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_faculty_department.R
\name{add_faculty_department}
\alias{add_faculty_department}
\title{Add Faculty Department}
\format{\describe{
 \item{department}{the faculty member's department}
 }}
\usage{
add_faculty_department(x)
}
\arguments{
\item{x}{data frame with raw.text column}
}
\value{
the input data frame along with a new columns.
}
\description{
The function accepts a data frame with raw.text column, and adds a column for the faculty's department
}
