% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_faculty_names.R
\name{add_faculty_names}
\alias{add_faculty_names}
\title{Add Faculty Names}
\format{\describe{
 \item{first.name}{Faculty's first name.}
 \item{last.name}{Faculty's last name.}
 }}
\usage{
add_faculty_names(x, complete = FALSE)
}
\arguments{
\item{x}{data frame with raw.text column}

\item{complete}{logical to indicate how many variables to add to x. Default is FALSE.}
}
\value{
the input data frame along with two new columns.
}
\description{
This function takes as input a data frame which includes the raw
  text associated with each faculty member. It returns that data frame along with
  two new columns: first.name and last.name. If \code{complete} is FALSE, it also
  returns full.name.
}
