% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_faculty_titles.R
\name{add_faculty_titles}
\alias{add_faculty_titles}
\title{Add Faculty Titles}
\format{\describe{
 \item{title}{Faculty's title}
 \item{rank}{Faculty's rank}
 }}
\usage{
add_faculty_titles(x)
}
\arguments{
\item{x}{data frame with raw.text column}
}
\value{
the input data frame along with a new columns.
}
\description{
This function takes as input a data frame which includes the raw
  text associated with each faculty member. It returns that data frame along with
  2 new columns: for the facutly's title and faculty's rank
}
