% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_gender.R
\name{add_gender}
\alias{add_gender}
\title{Add Gender}
\format{\describe{
  \item{gender}{gender as determined by the gender package.}
  }}
\usage{
add_gender(x)
}
\arguments{
\item{x}{data frame with \code{first.name} column}
}
\value{
the input data frame along with new column(s).
}
\description{
This function takes as input a data frame which includes the
  \code{first.name} and \code{birth.year} for each row. It returns that data frame along with a
  new column: \code{gender}.
}
