% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_graduate_gender.R
\name{add_graduate_gender}
\alias{add_graduate_gender}
\title{Add Graduate Gender}
\format{\describe{
  \item{gender}{Graduate's gender as determined by the gender package.}
  }}
\usage{
add_graduate_gender(x, complete = FALSE)
}
\arguments{
\item{x}{data frame with \code{first.name} column}

\item{complete}{logical to indicate how many variables to add to x. Default
is FALSE. If TRUE, other diagnostic variables are also added.}
}
\value{
the input data frame along with two new columns.
}
\description{
This function takes as input a data frame which includes the
  \code{first.name} of each graduate. It returns that data frame along with a
  new columns: \code{gender}.
}
