% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_graduate_honors.R
\name{add_graduate_honors}
\alias{add_graduate_honors}
\title{Add Graduate Honors}
\format{\describe{
    \item{honor}{details of distinction of primary thesis honor (if any)
                   We define the "primary" thesis honor as the first honor
                   appearing alongside grad's name}
    \item{major}{major in which graduate completed primary honor}
    \item{honor.2}{details of distinction of secondary thesis honor (if any)
                  We define the "secondary" thesis honor as the second honor
                 appearing alongside grad's name}
    \item{major.2}{major in which graduate completed secondary honor}
  }}
\usage{
add_graduate_honors(x)
}
\arguments{
\item{x}{data frame with raw.text column}
}
\value{
the input data frame along with four new columns.
}
\description{
This function takes as input a data frame which includes the raw
  text associated with each graduate. It returns that data frame along with
  four new columns associated with department honors.
}
