% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_graduate_names.R
\name{add_graduate_names}
\alias{add_graduate_names}
\title{Add Graduate Names}
\format{\describe{
 \item{first.name}{Graduate's first name.}
 \item{middle.name}{Graduate's middle name. Defined as the second separate string
       in full.name, if there are more than two names total.}
 \item{last.name}{Graduate's last name.}
 \item{full.name}{Graduate's entire name.}
 }}
\usage{
add_graduate_names(x)
}
\arguments{
\item{x}{data frame with raw.text column}
}
\value{
the input data frame along with three new columns.
}
\description{
This function takes as input a data frame which includes the raw
  text associated with each graduate. It returns that data frame along with
  four new columns: first.name, middle.name, last.name and full.name. We assume
  that the first/middle/last names consist of a single string. That is, even if
  someone's true last name is González del Riego, we record the last name as
  Riego.
}
