% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_graduate_race.R
\name{add_graduate_race}
\alias{add_graduate_race}
\title{Add Graduate Race}
\format{\describe{
  \item{race}{Graduate's race as determined by the wru package.}
  }}
\usage{
add_graduate_race(x, complete = FALSE)
}
\arguments{
\item{x}{data frame with \code{last.name} column}

\item{complete}{logical to indicate how many variables to add to x. Default
is FALSE. If TRUE, other diagnostic variables are also added.}
}
\value{
the input data frame along with new column(s).
}
\description{
This function takes as input a data frame which includes the
  \code{last.name} of each graduate. It returns that data frame along with a
  new column: \code{race}.
}
