% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_is_asian.R
\name{add_is_asian}
\alias{add_is_asian}
\title{Add Is Asian? Column}
\usage{
add_is_asian()
}
\value{
"graduates" dataframe with new column "is.asian" which is "TRUE" if we flag the lastname as asian,
        "FALSE" if we don't flag as asian (majority doesn't identify as Asian), or "NA" if we don't have
        data to make a guess.
}
\description{
This function loades in the "graduates" data, and adds a flag for asian lastnames.
             We flag a lastname as "asian" if a majority of people with the given lastname self-identified
             as Asian according to the U.S. Bureau of the Census
}
