% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_race.R
\name{add_race}
\alias{add_race}
\title{Add Graduate Race}
\format{\describe{
  \item{p_whi}{Prosterior probability that the individual's race is White, as determined by the wru package}
  \item{p_bla}{Prosterior probability that the individual's race is Black, as determined by the wru package}
  \item{p_asi}{Prosterior probability that the individual's race is Asian, as determined by the wru package}
  \item{p_his}{Prosterior probability that the individual's race is Hispanic, as determined by the wru package}
  \item{p_oth}{Prosterior probability that the individual's race is none of the above, as determined by the wru package}
  \item{race}{Graduate's race as the racial category with the highest prosterior probability.}
  }}
\usage{
add_race(x)
}
\arguments{
\item{x}{data frame with \code{last.name} column}
}
\value{
the input data frame along with new column(s).
}
\description{
This function takes as input a data frame which includes the
  \code{last.name} of each graduate. It returns that data frame along with a
  new column: \code{race}.
}
