% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_annual_dataframe.R
\name{build_annual_dataframe}
\alias{build_annual_dataframe}
\title{Build Annual Dataframe}
\format{\describe{
    \item{firstname}{firstname of graduate}
    \item{middlename}{middlename(s) of graduate}
    \item{lastname}{lastname of graduate}
    \item{year}{year of graduation}
    \item{latin.honor}{latin honor recieved by graduate (if any)}
    \item{honor}{details of distinction of primary thesis honor (if any)
                   We define the "primary" thesis honor as the first honor
                   appearing alongside grad's name}
    \item{major}{major in which graduate completed primary honor}
    \item{honor.2}{details of distinction of secondary thesis honor (if any)
                  We define the "secondary" thesis honor as the second honor
                 appearing alongside grad's name}
    \item{major.2}{major in which graduate completed secondary honor}
    \item{raw.text}{graduation details as appearing in course catalog}
    }}
\usage{
build_annual_dataframe(year)
}
\arguments{
\item{year}{year to build dataframe for}
}
\value{
a dataframe for the specified year of
                       # rows = # graduates for the year,
                       # variabels = 10
}
\description{
This function compiles information for the given year's degrees and graduates and presents
             it as a dataframe. Details on this process can be found in the vignette.
}
