% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_gender_df.R
\name{build_gender_df}
\alias{build_gender_df}
\title{Build Dataframe with Gender}
\usage{
build_gender_df()
}
\value{
new \code{graduates} dataframe with  \code{male_prob} and \code{female_prob} columns
}
\description{
This function adds \code{male_prob} and \code{female_prob} columns to the \code{graduate}
             dataset, and returns the new dataframe
}
