% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_annual_faculty.R
\name{create_annual_faculty}
\alias{create_annual_faculty}
\title{Create Current Faculty}
\usage{
create_annual_faculty(x)
}
\arguments{
\item{x}{the faculty.detail data set with year, leave, first.name, last.name, first.degree, first.degree.year, first.degree.school,
last.degree, last.degree.year, last.degree.school, birth.year, title, rank, status, department
gender, and race columns}
}
\value{
data set for faculty information for the last year
}
\description{
This function accepts the faculty.detail data frame, and returns information about the college's faculty for
             the current year (the last year for which information is available).
}
