% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_faculty_names.R
\name{fetch_lastname}
\alias{fetch_lastname}
\title{Fetch Lastname}
\usage{
fetch_lastname(name)
}
\arguments{
\item{name}{list of words in a name}
}
\value{
last name (as defined above)
}
\description{
Given a list of words in a name, this function will find the last name
             (defined as all words in the name after the first, not including collapsed middlenames if any).
}
