% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gather_faculty.R
\name{gather_faculty}
\alias{gather_faculty}
\title{Gather Faculty}
\format{\describe{
  \item{year}{year}
  \item{raw.text}{Raw text from the Course Catlog associated with each faculty member}
  \item{leave.full.year}{Boolean value indicating if faculty is on leave for academic year.}
  \item{leave.first.sem}{Boolean value indicating if faculty is on leave for first semester.}
  \item{leave.second.sem}{Boolean value indicating if faculty is on leave for second semester.}
  \item{leave.calender.year}{Boolean value indicating if faculty is on leave for calender year.}
  }}
\usage{
gather_faculty()
}
\value{
a dataframe for the specified years with a row for each faculty
  member and six variables.
}
\description{
This function brings together the information on all Williams
  faculty for the years that we have downloaded the data in the package. It
  also adds information about faculty on leave.
}
