% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gather_graduates.R
\name{gather_graduates}
\alias{gather_graduates}
\title{Gather Graduates}
\format{\describe{
  \item{year}{Graduation year}
  \item{raw.text}{Raw text from the Course Catlog associated with each graduate}
  \item{latin.honors}{Latin honors}
  \item{Phi.Beta.Kappa}{Boolean value indicating membership in Phi Beta Kappa}
  \item{Sigma.Xi}{Boolean value indicating membership in Sigma Xi}
  \item{birth.year}{Inferred birth year of graduate}
  }}
\usage{
gather_graduates()
}
\value{
a dataframe for the specified years with a row for each graduating
  senior and six variables.
}
\description{
This function brings together the information on all Williams
  graduates for the years that we have downloaded the data in the package. It
  also adds Latin honors information.
}
