% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_female_probability.R
\name{get_female_probability}
\alias{get_female_probability}
\title{Get Female Probability}
\usage{
get_female_probability(name, predict_gender)
}
\arguments{
\item{name}{name to predict gender for}

\item{predict_gender}{prediction table generated for all unique names in \code{graduates} dataset with
\code{gender} package}
}
\value{
probability that given name belongs to a female student. If information for given name is not
        present in \code{gender} package, we will return NA
}
\description{
This function returns the probability that a given name belongs to a female student according
             to the \code{gender} package.
}
