% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_asian.R
\name{is_asian}
\alias{is_asian}
\title{Is Asian?}
\usage{
is_asian(lastname, asian_lastnames)
}
\arguments{
\item{lastname}{lastname to flag}

\item{asian_lastnames}{data frame of the percentage of people with a given last name who identify as Asian}
}
\value{
TRUE if majority of people with given lastname self identify as ASIAN
        FALSE if majority of people with given lastname do not self identify as ASIAN
        NA if no data is available
}
\description{
Given a lastname, this function will flag it as "TRUE" (if asian), "FALSE" (if not asian),
             "NA" (if no data available). This is in accordance with data compiled by www.mongabay.com
             from the U.S. Bureau of the Census.
}
