% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_gender_predictions.R
\name{load_gender_predictions}
\alias{load_gender_predictions}
\title{Load Gender Predictions}
\usage{
load_gender_predictions()
}
\value{
A table containing gender prediction probabilities for each unique first name in the
        \code{graduates} dataset.
}
\description{
This function extracts the unique first names from the \code{graduates} dataset, and
             uses the \code{gender} package to predict the gender for each first name. The result is
             compiled as a table with relevant columns \code{proportion_female} and
             code{proportion_male} and returned to the calling function.
}
