% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_file.R
\name{read_file}
\alias{read_file}
\title{Read File}
\usage{
read_file(filepath)
}
\arguments{
\item{filepath}{filepath of file to read from}
}
\value{
contents of file in string format
}
\description{
reads in text from a file and returns it as a String
}
