% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_first_name.R
\name{scrape_first_name}
\alias{scrape_first_name}
\title{Scrape First Name}
\usage{
scrape_first_name(grad_details)
}
\arguments{
\item{grad_details}{Student's graduation details from catalog.}
}
\value{
student's firstname as infered by our definition: we define firstname as the first aphabetic sequence appearing in a student's
             name uptill (but not including) the first white space. For example, given graduation
             details, "DoHyun Tony Chung, with honors in Political Economy", we infer "DoHyun" as
             the student's firstname.
}
\description{
A student's graduation details is provided in the catalog in the following format:
             (1) <firstname> <midlename(s)> <lastname>, <theses details>
             or, (2) <firstname> <midlename(s)> <lastname>

             Here, we define firstname as the first aphabetic sequence appearing in a student's
             name uptill (but not including) the first white space. For example, given graduation
             details, "DoHyun Tony Chung, with honors in Political Economy", we infer "DoHyun" as
             the student's firstname.
}
