% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_last_name.R
\name{scrape_last_name}
\alias{scrape_last_name}
\title{Scrape Last Name}
\usage{
scrape_last_name(grad_details)
}
\arguments{
\item{grad_details}{Student's graduation details from catalog.}
}
\value{
student's lastname as infered by our definition:  we define lastname as the aphabetic sequence
       appearing in a student's name after (but not including) the last white space. For example,
       given graduation details, "DoHyun Tony Chung, with honors in Political Economy", we infer "Chung" as
       the student's lastname.
}
\description{
A student's graduation details is provided in the catalog in the following format:
             (1) <firstname> <midlename(s)> <lastname>, <theses details>
             or, (2) <firstname> <midlename(s)> <lastname>

             Here, we define lastname as the aphabetic sequence appearing in a student's
             name after (but not including) the last white space. For example, given graduation
             details, "DoHyun Tony Chung, with honors in Political Economy", we infer "Chung" as
             the student's lastname.
}
