% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_middle_name.R
\name{scrape_middle_name}
\alias{scrape_middle_name}
\title{Scrape Middle Name(s)}
\usage{
scrape_middle_name(grad_details)
}
\arguments{
\item{grad_details}{Student's graduation details from catalog.}
}
\value{
student's middle as infered by our definition:  we define middlename as the aphabetic sequence
       appearing in a student's name after (but not including) the first white space, and before
       (but not including) the last white space.
       For example, given graduation details, "DoHyun Tony Chung, with honors in Political Economy",
       we infer "Tony" as the student's middlename.
}
\description{
A student's graduation details is provided in the catalog in the following format:
             (1) <firstname> <midlename(s)> <lastname>, <theses details>
             or, (2) <firstname> <midlename(s)> <lastname>

             Here, we define middle as the aphabetic sequence appearing in a student's
             name after (but not including) the first white space, and before (but not including)
             the last white space.
             For example, given graduation details, "DoHyun Tony Chung, with honors in Political Economy",
             we infer "Tony" as the student's middle
}
