% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_primary_thesis_distinction.R
\name{scrape_primary_thesis_distinction}
\alias{scrape_primary_thesis_distinction}
\title{Scrape Primary Honor Distinction}
\usage{
scrape_primary_thesis_distinction(grad_details)
}
\arguments{
\item{grad_details}{A student's graduation details as appearing in the catalog.}
}
\value{
Student's primary honors distinction level (i.e. "HIGHEST", "HONORS", or "NONE")
}
\description{
A student's graduation details is provided in the catalog in the following format:
             (1) <firstname> <midlename(s)> <lastname>, <theses details>
             or, (2) <firstname> <midlename(s)> <lastname>

             This function concerns itself with the <theses details>. In turn, <theses details> are
             provided in the following format:
             <thesis 1 details>, <thesis 2 details>
             where each <thesis N detail> is in the format "with <distinction> honors in <major>"

             Here we define the first thesis as the "Primary Thesis". This function will scrape the
             distinction for the Primary Thesis which maybe one of the following: "HIGHEST" (for highest
             level of honors), "HONORS" (for thesis completion), "NONE" (if no thesis was written).
}
