% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_primary_thesis_major.R
\name{scrape_primary_thesis_major}
\alias{scrape_primary_thesis_major}
\title{Scrape Primary Thesis Major}
\usage{
scrape_primary_thesis_major(grad_details)
}
\arguments{
\item{grad_details}{A student's graduation details as appearing in the catalog.}
}
\value{
Student's primary thesis major
}
\description{
A student's graduation details is provided in the catalog in the following format:
             (1) <firstname> <midlename(s)> <lastname>, <theses details>
             or, (2) <firstname> <midlename(s)> <lastname>

             This function concerns itself with the <theses details>. In turn, <theses details> are
             provided in the following format:
             <thesis 1 details>, <thesis 2 details>
             where each <thesis N detail> is in the format "with <distinction> honors in <major>"

             Here we define the first thesis as the "Primary Thesis". This function will scrape the
             major for the Primary Thesis.
             For example, in the case of "DoHyun Tony Chung, with honors in Political Economy",
             we infer "Political Economy" as the student's primary thesis major.
}
