% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_secondary_thesis_major.R
\name{scrape_secondary_thesis_major}
\alias{scrape_secondary_thesis_major}
\title{Scrape Secondary Thesis Major}
\usage{
scrape_secondary_thesis_major(grad_details)
}
\arguments{
\item{grad_details}{A student's graduation details as appearing in the catalog.}
}
\value{
Student's secondary thesis major
}
\description{
A student's graduation details is provided in the catalog in the following format:
             (1) <firstname> <midlename(s)> <lastname>, <theses details>
             or, (2) <firstname> <midlename(s)> <lastname>

             This function concerns itself with the <theses details>. In turn, <theses details> are
             provided in the following format:
             <thesis 1 details>, <thesis 2 details>
             where each <thesis N detail> is in the format "with <distinction> honors in <major>"

             Here we define the second thesis as the "Secondary Thesis". This function will scrape the
             major for the Secondary Thesis.
             For example, in the case of "DoHyun Tony Chung, with honors in Political Economy, with
             honors in Mathematics", we infer "Mathematics" as the student's secondary thesis major.
}
