\name{setup.pdf.data}
\alias{setup.pdf.data}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Set Up the Williams College faculty pdf Data for Extraction
}
\description{
setup.pdf.data is the first step in extracting data from the pdf sourced
files. The function reads in the txt file generated by the pdf as
described in README.txt. The data is separated into the lines containing
professor names and titles, and the lines containing professors'
education information. The function returns a dataframe with those two
lines as its columns.
}
\usage{
setup.pdf.data(yr)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{yr}{
yr is the academic year of faculty data currently running through the
data extraction loop. yr must be numeric and expressed as a four-digit year
}
  }
\details{
There are minor discrepancies between the structure of the data in
academic year 2010 and all other years. These are taken care of in
setup.pdf.data such that future functions do not require year-specific
adjustments.
}
\value{
%%  ~Describe the value returned
%%  If it is a LIST, use
%%  \item{comp1 }{Description of 'comp1'}
%%  \item{comp2 }{Description of 'comp2'}
%% ...
}
\references{
%% ~put references to the literature/web site here ~
}
\author{
Andrew Leeser
}
\note{
%%  ~~further notes~~
}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
%% ~~objects to See Also as \code{\link{help}}, ~~~
}
\examples{
setup.pdf.data(current.year)
## Where current.year is the year variable being generated within the loop

setup.pdf.data(2010)
## To run the function for only academic year 2010
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
