% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize_graduation_details.R
\name{standardize_graduation_details}
\alias{standardize_graduation_details}
\title{Standardize Graduation Details}
\usage{
standardize_graduation_details(grad_details)
}
\arguments{
\item{grad_details}{A student's graduation details}
}
\value{
Standardizes deiscrepencies in graduation_details (if any)
}
\description{
This function standardizes a student's graduation details to account for discrepancies
             in the catalogs over the years. Each edge case is described in this description in detail.

            (1) "Women's, Gender and Sexuality", "Women's Gender & Sexuality" are treated as one
                 and replaced as "WGES".
}
