% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_gender_graduates.R
\name{add_gender_graduates}
\alias{add_gender_graduates}
\title{Add Gender Graduates}
\format{\describe{
  \item{gender}{character, gender as determined by the \code{gender} package.}
    \item{p_male.first}{numeric, probability that the individual is a male given their first name, as determined by the
    \code{gender} packege}
    \item{p_female.first}{numeric, probability that the individual is a female given their first name, as determined by the
    \code{gender} packege}
    \item{gender.first}{character,gender of an individual given their first name, as determined by the  \code{gender} packege}
    \item{p_male.middle}{numeric, probability that the individual is a male given their middle name, as determined by the
    \code{gender} packege}
    \item{p_female.middle}{numeric, probability that the individual is a female given their middle name, as determined by the
    \code{gender} packege}
    \item{gender.middle}{character,gender of an individual given their middle name, as determined by the  \code{gender} packege}
    \item{gender.g}{character, gender of an individual, as determined by \code{genderizeR} package}
  }}
\usage{
add_gender_graduates(x)
}
\arguments{
\item{x}{data frame with \code{birth.year}, \code{first.name} and
\code{middle.name} columns.}
}
\value{
the input data frame along with new column(s).
}
\description{
This function takes as input a data frame for graduates which includes the
  \code{first.name} and \code{birth.year} for each row. It returns that data
  frame along with new column \code{gender}, and other supporting
  information.
}
