% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_race.R
\name{add_race}
\alias{add_race}
\title{Add Graduate Race}
\format{\describe{
  \item{race}{Graduate's race as determined by the wru package.}
  }}
\usage{
add_race(x)
}
\arguments{
\item{x}{data frame with \code{last.name} column}
}
\value{
the input data frame along with new column(s).
}
\description{
This function takes as input a data frame which includes the
  \code{last.name} of each graduate. It returns that data frame along with a
  new column: \code{race}.
}
