% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_graduates.R
\name{create_graduates}
\alias{create_graduates}
\title{Create Data Frame of Graduates Data}
\format{\describe{
    \item{first.name}{First name of graduate}
    \item{last.name}{Last name of graduate}
    \item{year}{year of graduation}
    \item{latin.honor}{latin honor recieved by graduate (if any)}
    \item{honor}{details of distinction of primary thesis honor (if any)
                   We define the "primary" thesis honor as the first honor
                   appearing alongside grad's name}
    \item{major}{major in which graduate completed primary honor}
    \item{gender}{male or female, based on the \code{gender} package}
    }}
\usage{
create_graduates(complete = FALSE)
}
\arguments{
\item{complete}{logical to indicate how many variables to return in the data frame.
Default is FALSE, which returns the same information as in the \code{graduates}
data frame distributed with the package.}
}
\value{
a dataframe with a row for each graduating senior and 9 associated variables.
}
\description{
This function compiles information for all graduates and returns a
      dataframe.
}
\examples{

}
