% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gather_faculty.R
\name{gather_faculty}
\alias{gather_faculty}
\title{Gather Faculty}
\format{\describe{
  \item{year}{numeric <YYYY>, year}
  \item{raw.text}{character, Raw text from the Course Catlog 
      associated with each faculty member}
  \item{leave}{character, Faculty leave information; 
      one of "Academic Year", "First Semester", "Second Semester", 
      "Calendar Year", or "None"}
  }}
\usage{
gather_faculty()
}
\value{
a dataframe for the specified years with a row for each faculty
  member and six variables.
}
\description{
This function brings together the information on all Williams
  faculty for the years that we have downloaded the data in the package. It
  also adds information about faculty on leave.
}
