% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graduates.R
\docType{data}
\name{graduates}
\alias{graduates}
\title{Williams College Graduates}
\format{A tibble with 8,893 rows and 10 variables:
\describe{
    \item{year}{year of graduation}
    \item{first.name}{first name of graduate}
    \item{last.name}{last name of graduate}
    \item{gender}{based on \code{first.name}, using the gender package.}
    \item{race}{based on \code{last.name}, mainly using the wru package. Values are "Asian", "Black", "Hispanic", "Other", and "White"}
    \item{latin.honors}{character, Latin Honors recieved by graduate, if any}
    \item{Phi.Beta.Kappa}{logical, indicates Phi Beta Kappa membership}
    \item{Sigma.Xi}{logical, indicates Sigma Xi membership}
    \item{honor}{first department honor awarded, if any. A handful
      of graduates receive honors in two departments. We provide only the 
      first. Values are "honors" or "highest honors"}
    \item{major}{character, academic field in which graduate won primary honor,
       if any}
}}
\source{
\url{https://catalog.williams.edu/archive/}
}
\usage{
graduates
}
\description{
A tibble containing information for 8,893 graduates of Williams
College, classes of 2001 through 2016, extracted from the 
course catalog. Created by running \code{create_graduates()}.
}
\keyword{datasets}
