% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slice_and_demit.R
\name{slice_and_delimit}
\alias{slice_and_delimit}
\title{Slice and Delimit}
\usage{
slice_and_delimit(source_str, start_str, stop_str = "", delim = "\\n")
}
\arguments{
\item{source_str}{String containing target text}

\item{start_str}{String at which process should start}

\item{stop_str}{String at which process should end. The default is empty, which is equivalent to EOF.}

\item{delim}{character to delimite string on.}
}
\value{
target text in \code{source_str} delimited at character \code{delim}
}
\description{
This function slices the \code{source_str} at \code{start_str}
             till \code{stop_str}, and seperates the string in between as
             a list seperated by delimiter \code{delim}
}
