% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize_graduation_details.R
\name{standardize_graduation_details}
\alias{standardize_graduation_details}
\title{Standardize Graduation Details}
\usage{
standardize_graduation_details(grad_details)
}
\arguments{
\item{grad_details}{A student's graduation details}
}
\value{
Standardizes discrepencies in graduation_details (if any)
}
\description{
This function standardizes a student's graduation details to account for changes
             in the catalogs over the years.
}
